Separated preview generate logic from the component
