+++ src/pages/projContracts/api/docusign/downloadContract.ts
@@ -95,7 +95,6 @@ export const downloadContract = async (
   const { fileType, form  } = params;
   const { envDocFileKeys } = form;
 
-  console.log('download', params,  envDocFileKeys);
   if ( fileType === 'xlsx' || !envDocFileKeys.length) {
     return dlFromCocoServer(params);
   } else if (fileType === 'pdf' && envDocFileKeys.length) {
