+++ src/pages/projContracts/parts/Preview/Preview.tsx
@@ -1,96 +1,46 @@
 import {  Divider, Grid, Paper, Typography } from '@mui/material';
-import { useEffect, useState } from 'react';
-import { useSnackBar } from '../../../../hooks';
 import { Loading } from './Loading';
-
-import { downloadContract } from '../../api/docusign/downloadContract';
-import { TypeOfForm } from '../../form';
-import { base64ToBlob } from '../../../../lib';
 import { PreviewToolBar } from '../PreviewToolBar/PreviewToolBar';
-import { useFormikContext } from 'formik';
 import { DocumentsSelect } from './SelectDocuments';
 import { RefreshButton } from '../PreviewToolBar/RefreshButton';
 import { PreviewContainer } from './PreviewContainer';
+import { useContractPreview } from '../../hooks/useContractPreview';
 
 
 
 export const Preview = () => {
-  const { values, status } = useFormikContext<TypeOfForm>();
   const {
-    projId, projName, envelopeId,
-    envelopeStatus, envSelectedDoc,
-    revision,
-  } = values;
-  const [previewUrl, setPreviewUrl] = useState('');
-  const [previewLoading, setPreviewLoading] = useState(true);
-  const { setSnackState } = useSnackBar();
+    values,
+    previewLoading,
+    formLoading,
+    previewUrl,
+  } = useContractPreview();
 
-  const handlePreview = async () => {
-    try {
-      setPreviewLoading(true);
-      const res = await downloadContract({
-        form: values,
-        fileType: 'pdf',
-      });
-
-      if (!res) return;
-      if (previewUrl) URL.revokeObjectURL(previewUrl); // free Memory
-
-      const base64 = res;
-
-      if (base64) {
-        const blob = base64ToBlob( base64, 'application/pdf' );
-        const url = URL.createObjectURL( blob );
-        setPreviewUrl(url);
-      } else {
-        setPreviewUrl('');
-      }
-
-    } catch (err) {
-
-      setSnackState({
-        open: true,
-        severity: 'error',
-        message: `プレビューの取得が失敗しました。管理者をご連絡ください。${err.message}`,
-      });
-
-    } finally {
-      setPreviewLoading(false);
-    }
-
-  };
-
-  useEffect(()=>{
-
-    if (!projId || !projName || (status as TFormStatus) === 'busy') return;
-
-    handlePreview();
-  }, [
-    projId,
-    projName,
-    envelopeStatus,
-    revision,
-    envSelectedDoc,
-    status,
-  ]);
+  const { envelopeId, envelopeStatus, projId, projName } = values;
 
 
 
-  const loading = (status as TFormStatus) === 'busy' || previewLoading;
 
 
   return (
     <PreviewContainer>
       <Grid item xs={6}>
-        <RefreshButton loading={loading} isVisible={!!projId} />
+        <RefreshButton loading={formLoading} isVisible={!!projId} />
       </Grid>
       <Grid item xs={6}>
-        <PreviewToolBar {...{ envelopeId, envelopeStatus, loading, projId, projName, previewLoading }} />
+        <PreviewToolBar {...{
+          envelopeId,
+          envelopeStatus,
+          loading: formLoading,
+          projId,
+          projName,
+          previewLoading }}
+        />
       </Grid>
       <Grid item xs={12}>
         <Divider />
       </Grid>
-      {!loading && previewUrl &&
+      {!formLoading && previewUrl &&
       <Grid item xs={12}>
         <Paper>
           <DocumentsSelect />
@@ -99,7 +49,7 @@ export const Preview = () => {
         </Paper>
       </Grid>}
 
-      {loading && projId &&
+      {formLoading && projId &&
       <Grid item xs={12}>
         <Loading />
         <Typography variant="caption">
@@ -107,7 +57,7 @@ export const Preview = () => {
         </Typography>
       </Grid>}
 
-      {loading && !projId &&
+      {formLoading && !projId &&
       <Grid item xs={12}>
         <Typography variant="caption">
           プロジェクトを選択してください。
